Fix 64bit error in SipMessage.cpp
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@18553 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXtackLib/src/net/SipMessage.cpp
@@ -3509,7 +3509,7 @@ UtlBoolean SipMessage::getCSeqField(int* sequenceNum, UtlString* sequenceMethod)
                           "or it was followed by an unexpected character.  "
                           "Problem was detected after %d characters were processed.",
                           &value[valueStart],
-                          endptr - &value[valueStart]);
+                          (int)(endptr - &value[valueStart]));
          }
       }
    }
